Conflicts:
	BobAndFriends/BobAndFriends/bin/Debug/BobAndFriends.vshost.application
	BobAndFriends/BobAndFriends/bin/Debug/BobAndFriends.vshost.exe.manifest

+++ Crawler/src/Boris.py
@@ -3,10 +3,10 @@ import glob
 from ConfigParser import SafeConfigParser
 from threading import Thread
 
-from Crawlers import TradeTracker, FeedCrawler, Effiliation, JacobElektronik
+from Crawlers import Effiliation, JacobElektronik, Amazon, TradeTracker, FeedCrawler
 from Crawlers import Belboon
 from Crawlers import Daisycon
-from Crawlers import Bol
+#from Crawlers import Bol
 from Crawlers import LDLC
 from Crawlers import Linkshare
 from Crawlers import PepperjamNetwork
@@ -19,7 +19,6 @@ def main():
     parseConfigFile()
     emptyDirectories()
     startCrawlers()
-
 
 # Procedure to parse the config file
 def parseConfigFile():
@@ -35,10 +34,18 @@ def emptyDirectories():
     productFeedPaths = [x[0] for x in os.walk(productFeedsPath)]
 
     for path in productFeedPaths:
-        if path != productFeedPaths[0] and 'BorderBot' not in path and "Rene's Toppertjes" not in path:
+        if path != productFeedPaths[0] and 'BorderBot' not in path and "Rene's Toppertjes" not in path and 'PepperjamNetwork' not in path:
             files = glob.glob(path + '/*')
-            for f in files:
-                os.remove(f)
+
+            while True:
+                try:
+                    for f in files:
+                        os.remove(f)
+                    break
+                except WindowsError as e:
+                    print e
+                    raw_input("The above product feed is opened by a program, possibly Sublime. "
+                              "Close the feed and press any key to continue")
 
 
 # This procedure starts all crawlers.
@@ -55,8 +62,8 @@ def startCrawlers():
     t = Thread(target=Daisycon.Crawler().main)
     t.start()
 
-    t = Thread(target=Bol.Crawler().main)
-    t.start()
+    #t = Thread(target=Bol.Crawler().main)
+    #t.start()
 
     t = Thread(target=Effiliation.Crawler().main)
     t.start()
@@ -72,5 +79,9 @@ def startCrawlers():
 
     t = Thread(target=PepperjamNetwork.Crawler().main)
     t.start()
+
+    t = Thread(target=Amazon.Crawler().main)
+    t.start()
+
 
 main()
